// submodules/common-dto/api-client/main/models/CityDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CityDto = {
    /**
     * Id of city
     */
    id: number;
    /**
     * Localized title of city
     */
    title: string;
    /**
     * Localized title of country
     */
    country: string;
};