// submodules/common-dto/api-client/main/models/MainAddressPostRequestDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MainAddressPostRequestDto = {
    /**
     * Address owner username
     */
    username?: string;
    /**
     * Country id
     */
    country_id: number;
    /**
     * City name
     */
    city: string;
    /**
     * Street name
     */
    street: string;
    /**
     * House number
     */
    house_number?: string;
    /**
     * Postal code
     */
    zipcode: string;
    /**
     * Full name
     */
    fullname: string;
    /**
     * Phone number
     */
    phone: string;
};