// submodules/common-dto/api-client/main/models/MainAddressPutIdRequestDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MainAddressPutIdRequestDto = {
    /**
     * Country id
     */
    country_id?: number;
    /**
     * City name
     */
    city?: string;
    /**
     * Street name
     */
    street?: string;
    /**
     * House number
     */
    house_number?: string | null;
    /**
     * Postal code
     */
    zipcode?: string;
    /**
     * Full name
     */
    fullname?: string;
    /**
     * Phone number
     */
    phone?: string;
    /**
     * Is active
     */
    is_active?: boolean;
};